supports variable event types
