+++ website/app/components/doc/link-with-icon/index.hbs
@@ -9,7 +9,7 @@
     @route={{@route}}
     @models={{this.models}}
     ...attributes
-    {{doc-track-event @eventName}}
+    {{doc-track-event-on "click" @eventName}}
   >
     {{@label}}
     <FlightIcon class="doc-link-with-icon__icon" @name={{@icon}} />
@@ -21,7 +21,7 @@
     href={{@href}}
     target="_blank"
     rel="noopener noreferrer"
-    {{doc-track-event @eventName}}
+    {{doc-track-event-on "click" @eventName}}
     ...attributes
   >
     {{@label}}
